+++ popup/js/__tests__/initBookmarkManager.test.js
@@ -177,7 +177,7 @@ describe('initBookmarkManager cleanup apply', () => {
       __esModule: true,
       createExtensionContext: () => ext,
     }))
-    await jest.unstable_mockModule('../model/options.js', () => ({
+    await jest.unstable_mockModule('../model/optionsStorage.js', () => ({
       __esModule: true,
       defaultOptions: {},
       getEffectiveOptions: jest.fn(() => Promise.resolve({})),